+++ src/App.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import { QueryProvider } from './app/providers/QueryProvider';
 import { AuthProvider } from './app/providers/AuthProvider';
+import { ToastProvider } from './app/providers/ToastProvider';
 import { AppRouter } from './app/router';
 
 function App() {
   return (
     <QueryProvider>
-      <AuthProvider>
-        <AppRouter />
-      </AuthProvider>
+      <ToastProvider>
+        <AuthProvider>
+          <AppRouter />
+        </AuthProvider>
+      </ToastProvider>
     </QueryProvider>
   );
 }